+++ tests/Unit/Panels/CitiesPanelTest.php
@@ -2,9 +2,8 @@
 
 namespace Wearesho\GoogleAutocomplete\Yii\Tests\Unit\Panels;
 
-use GuzzleHttp;
 use Wearesho\GoogleAutocomplete\Yii\Panels\CitiesPanel;
-use Wearesho\GoogleAutocomplete\Yii\Tests\TestCase;
+use Wearesho\GoogleAutocomplete\Yii\Tests\PanelTestCase;
 
 /**
  * Class CitiesPanelTest
@@ -12,81 +11,46 @@ use Wearesho\GoogleAutocomplete\Yii\Tests\TestCase;
  * @coversDefaultClass CitiesPanel
  * @internal
  */
-class CitiesPanelTest extends TestCase
+class CitiesPanelTest extends PanelTestCase
 {
-    /** @var CitiesPanel */
-    protected $panel;
+    protected const TESTED_PANEL = CitiesPanel::class;
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->panel = CitiesPanel::instance();
-    }
-
-    public function testRulesArray(): void
-    {
-        $this->assertEquals(
+        $this->expectedRules = [
             [
-                [
-                    ['token', 'input',],
-                    'required',
-                ],
-                [
-                    ['token', 'input', 'language'],
-                    'string',
-                ],
-                ['language', 'string', 'length' => 2,],
-                [
-                    'language',
-                    'default',
-                    'value' => function (): string {
-                        return \Yii::$app->language;
-                    },
-                ],
+                ['token', 'input',],
+                'required',
             ],
-            $this->panel->rules()
-        );
-    }
-
-    public function testValidate(): void
-    {
-        $this->setQueryAttributes([
-            'input' => 'input',
-            'token' => $this->token
-        ]);
-        $this->assertTrue($this->panel->validate());
+            [
+                ['token', 'input', 'language'],
+                'string',
+            ],
+            ['language', 'string', 'length' => 2,],
+            [
+                'language',
+                'default',
+                'value' => function (): string {
+                    return \Yii::$app->language;
+                },
+            ],
+        ];
     }
 
     public function testFullResponse(): void
     {
-        $stack = GuzzleHttp\HandlerStack::create(new GuzzleHttp\Handler\MockHandler([
-            new GuzzleHttp\Psr7\Response(200, [], json_encode([
-                'status' => 'OK',
-                'predictions' => [
-                    [
-                        'description' => 'description',
-                        'structured_formatting' => [
-                            'main_text' => 'needle',
-                        ],
-                        'terms' => []
-                    ],
-                    [],
-                ],
-            ])),
-        ]));
-        $client = new GuzzleHttp\Client(['handler' => $stack]);
-        \Yii::$container->set(GuzzleHttp\ClientInterface::class, $client);
+        $this->mock($this->getSuccessResponse());
         $this->setQueryAttributes([
             'input' => 'input',
             'token' => $this->token,
             'city' => 'city',
             'type' => 'type',
-            'language' => 'uk',
         ]);
-        $this->panel = \Yii::$container->get(CitiesPanel::class);
+        \Yii::$app->language = 'uk';
 
-        $this->assertTrue($this->panel->validate());
-        $this->assertEquals('needle', $this->panel->getResponse()->data[0]);
+        $this->assertEquals('needle', $this->instancePanel()->getResponse()->data[0]);
     }
 }
