+++ tests/Unit/Panels/StreetsPanelTest.php
@@ -4,8 +4,7 @@ namespace Wearesho\GoogleAutocomplete\Yii\Tests\Unit\Panels;
 
 use GuzzleHttp;
 use Wearesho\GoogleAutocomplete\Yii\Panels\StreetsPanel;
-use Wearesho\GoogleAutocomplete\Yii\Tests\TestCase;
-use yii\web\HttpException;
+use Wearesho\GoogleAutocomplete\Yii\Tests\PanelTestCase;
 
 /**
  * Class StreetsPanelTest
@@ -13,8 +12,10 @@ use yii\web\HttpException;
  * @coversDefaultClass StreetsPanel
  * @internal
  */
-class StreetsPanelTest extends TestCase
+class StreetsPanelTest extends PanelTestCase
 {
+    protected const TESTED_PANEL = StreetsPanel::class;
+
     /** @var StreetsPanel */
     protected $panel;
 
@@ -23,66 +24,34 @@ class StreetsPanelTest extends TestCase
         parent::setUp();
 
         $this->panel = StreetsPanel::instance();
-    }
-
-    public function testRulesArray(): void
-    {
-        $this->assertEquals(
+        $this->expectedRules = [
             [
-                [
-                    ['token', 'input',],
-                    'required',
-                ],
-                [
-                    ['token', 'input', 'language'],
-                    'string',
-                ],
-                ['language', 'string', 'length' => 2,],
-                [
-                    'language',
-                    'default',
-                    'value' => function (): string {
-                        return \Yii::$app->language;
-                    },
-                ],
-                [
-                    ['type', 'city',],
-                    'string'
-                ],
-                ['type', 'filter', 'filter' => 'preg_quote'],
+                ['token', 'input',],
+                'required',
             ],
-            $this->panel->rules()
-        );
-    }
-
-    public function testValidate(): void
-    {
-        $this->setQueryAttributes([
-            'input' => 'input',
-            'token' => $this->token
-        ]);
-        $this->assertTrue($this->panel->validate());
+            [
+                ['token', 'input', 'language'],
+                'string',
+            ],
+            ['language', 'string', 'length' => 2,],
+            [
+                'language',
+                'default',
+                'value' => function (): string {
+                    return \Yii::$app->language;
+                },
+            ],
+            [
+                ['type', 'city',],
+                'string'
+            ],
+            ['type', 'filter', 'filter' => 'preg_quote'],
+        ];
     }
 
     public function testFullResponse(): void
     {
-        $stack = GuzzleHttp\HandlerStack::create(new GuzzleHttp\Handler\MockHandler([
-            new GuzzleHttp\Psr7\Response(200, [], json_encode([
-                'status' => 'OK',
-                'predictions' => [
-                    [
-                        'description' => 'description',
-                        'structured_formatting' => [
-                            'main_text' => 'needle',
-                        ],
-                        'terms' => []
-                    ],
-                    [],
-                ],
-            ])),
-        ]));
-        $client = new GuzzleHttp\Client(['handler' => $stack]);
-        \Yii::$container->set(GuzzleHttp\ClientInterface::class, $client);
+        $this->mock($this->getSuccessResponse());
         $this->setQueryAttributes([
             'input' => 'input',
             'token' => $this->token,
@@ -90,62 +59,36 @@ class StreetsPanelTest extends TestCase
             'type' => 'type',
             'language' => 'uk',
         ]);
-        $this->panel = \Yii::$container->get(StreetsPanel::class);
-
-        $this->assertTrue($this->panel->validate());
-        $this->assertEquals('needle', $this->panel->getResponse()->data[0]);
+        $this->assertEquals('needle', $this->instancePanel()->getResponse()->data[0]);
     }
 
     public function testShortResponse(): void
     {
-        $stack = GuzzleHttp\HandlerStack::create(new GuzzleHttp\Handler\MockHandler([
-            new GuzzleHttp\Psr7\Response(200, [], json_encode([
-                'status' => 'OK',
-                'predictions' => [
-                    [
-                        'description' => 'description',
-                        'structured_formatting' => [
-                            'main_text' => 'needle',
-                        ],
-                        'terms' => []
-                    ],
-                    [],
-                ],
-            ])),
-        ]));
-        $client = new GuzzleHttp\Client(['handler' => $stack]);
-        \Yii::$container->set(GuzzleHttp\ClientInterface::class, $client);
+        $this->mock($this->getSuccessResponse());
         $this->setQueryAttributes([
             'input' => 'input',
             'token' => $this->token,
             'language' => 'uk',
         ]);
-        unset($_GET['type']);
-        unset($_GET['city']);
-        $this->panel = \Yii::$container->get(StreetsPanel::class);
-
-        $this->assertTrue($this->panel->validate());
-        $this->assertEquals('needle', $this->panel->getResponse()->data[0]);
+        $this->unsetQueryAttributes(['type', 'city']);
+        $this->assertEquals('needle', $this->instancePanel()->getResponse()->data[0]);
     }
 
+    /**
+     * @expectedException \yii\web\HttpException
+     */
     public function testUnavailableService(): void
     {
-        $stack = GuzzleHttp\HandlerStack::create(new GuzzleHttp\Handler\MockHandler([
-            new GuzzleHttp\Exception\RequestException(
-                'Some exception',
-                new GuzzleHttp\Psr7\Request('GET', 'google.com', [])
-            )
-        ]));
-        $client = new GuzzleHttp\Client(['handler' => $stack]);
-        \Yii::$container->set(GuzzleHttp\ClientInterface::class, $client);
+        $this->mock(new GuzzleHttp\Exception\RequestException(
+            'Some exception',
+            new GuzzleHttp\Psr7\Request('GET', 'google.com', [])
+        ));
         $this->setQueryAttributes([
             'input' => 'input',
             'token' => $this->token,
             'language' => 'uk',
         ]);
-        $this->panel = \Yii::$container->get(StreetsPanel::class);
 
-        $this->expectException(HttpException::class);
-        $this->panel->getResponse();
+        $this->instancePanel()->getResponse();
     }
 }
